Removed redundant tdd folder and implemented test case 1
