+++ tdd_lab/src/counter.py
@@ -4,13 +4,13 @@ Counter API Implementation
 from flask import Flask, jsonify
 from . import status
 
-COUNTERS = {}
-
 app = Flask(__name__)
 
 def counter_exists(name):
-    """Check if counter exists"""
-    return name in COUNTERS
+  """Check if counter exists"""
+  return name in COUNTERS
+
+COUNTERS = {}
 
 @app.route('/counters/<name>', methods=['POST'])
 def create_counter(name):
@@ -20,12 +20,12 @@ def create_counter(name):
     COUNTERS[name] = 0
     return jsonify({name: COUNTERS[name]}), status.HTTP_201_CREATED
 
-MINECRAFT_COUNTERS = {}
+Minecraft_Coutner = {}
 
-@app.route('/minecraft_server/<name>', methods=['POST'])
+@app.route('/minecraft/<name>', methods=['POST'])
 def create_minecraft_counter(name):
     """Create a counter"""
     if counter_exists(name):
-        return jsonify({"error": f"Minecraft counter {name} already exists"}), status.HTTP_409_CONFLICT
+        return jsonify({"error": f"Counter {name} already exists"}), status.HTTP_409_CONFLICT
     COUNTERS[name] = 0
-    return jsonify({name: MINECRAFT_COUNTERS[name]}), status.HTTP_201_CREATED
+    return jsonify({name: Minecraft_Coutner[name]}), status.HTTP_201_CREATED
