--- tdd_lab/tdd/pytest.ini
@@ -1,10 +1,0 @@
-[pytest]
-addopts = --verbose --cov=src --cov-report=term-missing
-
-[coverage:run]
-omit =
-    */tests/*
-    */migrations/*
-
-[coverage:report]
-show_missing = True
