--- tdd_lab/tdd/requirements.txt
@@ -1,3 +1,0 @@
-pytest
-pytest-cov
-Flask
