--- tdd_lab/tdd/src/__init__.py
@@ -1,7 +1,0 @@
-import pytest
-
-# we need to import the unit under test - counter
-from src.counter import app 
-
-# we need to import the file that contains the status codes
-from src import status
