--- tdd_lab/tdd/src/status.py
@@ -1,7 +1,0 @@
-# HTTP status codes
-HTTP_200_OK = 200
-HTTP_201_CREATED = 201
-HTTP_204_NO_CONTENT = 204
-HTTP_404_NOT_FOUND = 404
-HTTP_405_METHOD_NOT_ALLOWED = 405
-HTTP_409_CONFLICT = 409
