+++ tdd_lab/tests/test_counter.py
@@ -10,3 +10,26 @@ how to call the web service and assert what it should return.
 - The service must be able to update a counter by name.
 - The service must be able to read the counter
 """
+import pytest
+from src import app
+from src import status
+
+@pytest.fixture()
+def client():
+    """Fixture for Flask test client"""
+    return app.test_client()
+
+@pytest.mark.usefixtures("client")
+class TestCounterEndpoints:
+    """Test cases for Counter API"""
+
+    def test_create_counter(self, client):
+        """It should create a counter"""
+        result = client.post('/counters/foo')
+        assert result.status_code == status.HTTP_201_CREATED
+
+    def minecraft_counter(self, client):
+        # Creates a counter on the minecraft endpoint
+
+        result = client.post('/minecraft/steve')
+        assert result.status_code == status.HTTP_201_CREATED
